make package lib for websocket
